requirements.txt
PyGithub
oci
oracledb
pyyaml
requests
nltk
python-dotenv
sentence-transformers

# Legacy (kept for backward compatibility with old pipeline)
scrapy
